+++ src/app/api/subscribe/route.ts
@@ -1,6 +1,6 @@
 import { NextResponse } from 'next/server';
 import * as brevo from '@getbrevo/brevo';
-import { EMAIL_REGEX } from '@/constants';
+import { EMAIL_REGEX } from '@/validation';
 
 interface BrevoApiError {
   response?: {
